First pass at the bulk of the DNA function
This might work, I haven't tested it yet. It creates a ship DNA string and
creates a link from the top of an EFT block to open the fitting in game. It'll
also fix line breaks so users don't have to add '<br />' themselves.

Or at least, that's what I think I wrote. It may do something else entirely.