Refactor to use inheritance for the typeID mapping
I'm planning on adding a way to query a database to map the names to type IDs, but for now I'm just using the gigantic array. To make it easy to add more functionality in the future, the mapping strategy is configurable, and the only thing that should need to be done is extend EveFittingEFTParser and add a block to the EveFittingSetup function in EveFitting.php.